refactor: move the session list columns out of mutations.go
mutations.go had grown past the 800 line ceiling. The list column trio
moves to mcpservers.go, where mcpservers_test.go had been sitting with no
source file beside it.

TestDecodeMCPServersRefusesWhatItCannotRead is renamed: decodeStrings is
shared by both list columns now, so the old name no longer said what the
test pins. The assertions are untouched.
